+++ aci/common/src/main/java/com/yanny/aci/tooltip/CoreTooltipUtils.java
@@ -9,20 +9,34 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class CoreTooltipUtils {
+    /** @deprecated use {@link #toComponents(List, int, boolean, TooltipStyle)} */
+    @Deprecated(forRemoval = true, since = "1.1.0")
     @NotNull
     public static List<Component> toComponents(List<TooltipNode> tooltip, int pad, boolean showAdvancedTooltip) {
+        return toComponents(tooltip, pad, showAdvancedTooltip, TooltipStyle.DEFAULT);
+    }
+
+    @NotNull
+    public static List<Component> toComponents(List<TooltipNode> tooltip, int pad, boolean showAdvancedTooltip, TooltipStyle style) {
         List<Component> components = new ArrayList<>();
 
         for (TooltipNode node : tooltip) {
-            components.addAll(toComponents(node, pad, showAdvancedTooltip));
+            components.addAll(toComponents(node, pad, showAdvancedTooltip, style));
         }
 
         return components;
     }
 
+    /** @deprecated use {@link #toComponents(TooltipNode, int, boolean, TooltipStyle)} */
+    @Deprecated(forRemoval = true, since = "1.1.0")
     @NotNull
     public static List<Component> toComponents(TooltipNode tooltip, int pad, boolean showAdvancedTooltip) {
-        return tooltip.getComponents(pad, showAdvancedTooltip);
+        return toComponents(tooltip, pad, showAdvancedTooltip, TooltipStyle.DEFAULT);
+    }
+
+    @NotNull
+    public static List<Component> toComponents(TooltipNode tooltip, int pad, boolean showAdvancedTooltip, TooltipStyle style) {
+        return tooltip.getComponents(pad, showAdvancedTooltip, style);
     }
 
     @NotNull